6月11日 1、增加了中文名的应用设置 2、区分了两个应用的图标 3、修改了意见反馈中的progressView退出不会消失的bug 4、修改了查找班级按键显示错误的bug 5、修改了学生版的info.plist文件的名字

+++ ChuangYeBa/TestViewController.m
@@ -286,10 +286,9 @@ static NSString *testStateCellIdentifier = @"TestStateCell";
 - (TestStateCell *)tableView:(UITableView *)tableView testStateCellForRowAtIndexPath:(NSIndexPath *)indexPath {
     TestStateCell *testStateCell = [tableView dequeueReusableCellWithIdentifier:testStateCellIdentifier];
     testStateCell.typeLabel.text = @"单项选择练习";
-    testStateCell.currentLabel.text = [NSString stringWithFormat:@"%lu", quizNo];
-    testStateCell.totalLabel.text = [NSString stringWithFormat:@"%lu", self.quizs.count];
+    testStateCell.currentLabel.text = [NSString stringWithFormat:@"%ld", quizNo];
+    testStateCell.totalLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)self.quizs.count];
     return testStateCell;
-
 }
 
 - (QuestionCell *)tableView:(UITableView *)tableView questionCellForRowAtIndexPath:(NSIndexPath *)indexPath {
